Split registries (#30)
* split forms and views (registries)

* add element imports to respective registries

* remove main and require importing elements explicitly

* update snippets

* move element code up in the tree

* extract boilerplate to factory

* update readme
