+++ lit-any/forms/TemplateRegistry.js
@@ -1,0 +1,8 @@
+import TemplateRegistry from '../template-registry';
+
+export default class FormTemplateRegistry extends TemplateRegistry {
+    useComponents(components) {
+        this.components = components;
+        return this;
+    }
+}
