--- lit-any/index.js
@@ -1,10 +1,0 @@
-import './elements/lit-view';
-import './elements/lit-form';
-import { ViewTemplates, FieldTemplates } from './template-registry';
-import render from './render';
-
-export {
-    ViewTemplates,
-    FieldTemplates,
-    render,
-};
