+++ storybook/notes/components/polymer-elements/dropdown.js
@@ -44,9 +44,9 @@ dropdown({
 
 --- js
 import { dropdown } from '@lit-any/components-paper-elements';
-import { FieldTemplates } from 'lit-any;
+import Registry from 'lit-any/forms;
 
-FieldTemplates.default
+Registry.default
     .when(f => f.type === 'string')
     .renders(dropdown(options));
 ---
