+++ storybook/notes/components/vaadin/textbox.js
@@ -16,9 +16,9 @@ ${field}
 
 --- js
 import { textbox } from '@lit-any/components-vaadin';
-import { FieldTemplates } from 'lit-any;
+import Registry from 'lit-any/forms;
 
-FieldTemplates.default
+Registry.default
     .when(f => f.type === 'string')
     .renders(textbox(options));
 ---
